Sort the tracks after search
This change is a strange hack which I do not understand. But it takes
into account two things:

* The HTTPMS server returns the tracks properly sorted already
* Rhythmbox displays them in reverse order of adding them to the source
db

Try as hard as I might, I wasn't able to change the sorting behaviour of
the source. The GTK and Rhythmbox documentation is lacking. Hm... no.
Rather, it is completely missing. One have to deal with the API
reference and guess for himself.