Sort different instances of the same album separately
An album might be in the library more than once. For example, one might
have a flac and and mp3 version. Or something. In this case Rhythmbox
naturally sorts the by track number alone sucha as:

    1. Track 1
    1. Track 1
    2. Track 2
    2. Track 2

Where the first column is the track number an the rest is the track
name.

This commit makes it so that the mutliple versions of the same album are
sorted only between themsevels. Which means the example above will
become:

    1. Track 1
    2. Track 2
    1. Track 1
    2. Track 2

I think this makes much more sense!